requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cssselect==1.2.0
cycler==0.11.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==2.3.3
Flask-MySQLdb==2.0.0
fonttools==4.42.0
Frozen-Flask==0.18
google-cloud-logging
google-cloud-storage
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mysqlclient==2.2.0
networkx==3.1
numpy==1.25.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.15.0
pycparser==2.21
pyparsing==3.0.9
pyquery==2.0.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-louvain==0.16
pytz==2023.3
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tenacity==8.2.2
threadpoolctl==3.3.0
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.7
wsproto==1.2.0
xmltodict==0.13.0
